Show heatmap options when heatmap is run
Heatmap options are now shown after the user clicks the button

I also reordered the Heatmap and Partial Correlation buttons, since it's
awkward to get the Heatmap options to display well if its button is
second. We'll probably completely rework this later, but for now this
looks a little less ugly.