Add some search tests to check some fulltext searches
The tests just check that the correct number of results are returned. If
there's someething wrong with the fulltext searches, it should cause an
issue with at least one of these tests.

I'm not sure if this is the right place to put these tests, so they might need to be moved somewhere else later.