Remove double-encoding to json
Passing the data into `requests.post` as a `json=...` argument will
automatically encode the data to JSON and set up the correct
Content-Type header.

The call `json.dumps(post_data)` was pre-encoding the data to a JSON
string, that was the re-encoded to JSON yet again, which is not what
we want. This commit fixes that.X