Fix the way database_connector is called
The function returns a call to mdb.connect, which only returns the
connection object. It was throwing an error due to trying to unpack
multiple items.

One thing I noticed while looking at this is that we should probably
decide on a single way to create these connections. In some other files,
we instead use database_connection from wqflask.database (instead of the
gn3 database_connector from gn3.db_utils)