Fix combined searches involving aliases
The logic for building the SQL query wasn't working correctly when
aliases were involved. The query also checks for a gene symbol's aliases
in the DB, but this wasn't properly handling the AND/OR logic, leading
to no results being shown when doing a "Get Combined" search involving a
gene symbol.

This also removed some unused code related to aliases.