Fix issue with collection page export
There were some extra spaces/newline characters in the checkbox table
cell that were causing the javascript fetching trait info to not work
correctly (it was trying to select the first item in the table cell,
which in this case was the space/newline characters instead of the
checkbox itself). Removing those characters seem to fix the export.