Refactor: pass redis and db connections as argument
Pass the redis and database connection/cursors as arguments to the
class methods, and do not retain a copy of the connections/cursors.

This allows us to do the connection management in the context
managers elsewhere - ideally, at the top-level. For now the context
manager is within the `create_dataset` function, but this should be
moved out to a higher level once the lower levels are verified to be
working as expected.