Make some changes to prevent show_corr_results.py bugs (details below)
At one point the MonadicDict of the dataset ob was being passed to create_trait, which does not work. I changed it back to passing the object

Another place calls "as_dict" on the dataset ob, which doesn't work; I had to call as_monadic_dict().data for that

Some traits are also None for some reason, which caused an error in the filter. For now I just used an if/else, but someone can deal with this differently later