oauth2: Migrate user details and data
Migrate the user details when the user successfully logs in using the
old system. This helps "copy" the user details over to the new auth
database.

Once the user details have been "copied" over from the older auth
system (stored in Redis), the resources data attached to each user can
then be migrated over for the user once they log in to the newer auth
system.