oauth2: Load menu and populate dropdown
In order to enable searching through the existing datasets for linking
with the auth system, we need a way to filter the data.

In this commit, I load the menu and load it onto a dropdown box as a
proof of concept to give an idea of some of the filters I can use down
the road to fetch the data.

This commit breaks the data-linking code. I will need to rework the
data-linking code entirely to get it working again.