Fix main search bug causing AND searches to no longer work
The cause of this bug is that, when the autocomplete feature was
implemented, it checks the OR search field (in order to save its
contents)  and only submits the form if
it isn't empty. This means that, if the user fills out the AND field and
submits the form, nothing happens.

I changed it to check both the AND and OR fields (and save both of their
contents). While not perfect (since only one field is ever actually
used), I figured it's best to just store everything (and there isn't
really any way to know which search is intended when both fields are
filled).