(Hopefully)fix bug where the mapping form submission is sometimes wrongly treated as a GET, causing the code to check for a hash_of_inputs input that doesn't exist.
This is hard to troubleshoot because I could never reproduce the issue
(only Rob has mentioned it), but the error Rob showed me seems to
directly imply this was issue, so hopefully this change (which directly
checks if hash_of_inputs is set) will prevent it
from happening. It's possible it will still throw a different error,
though (if it's actually treating it as a GET, it presumably won't have
any of the other inputs it expects). This issue is a bit confusing,
because I'm not sure how/why mapping run from the trait page would ever
be treated as a GET request (since the show_trait page form is POST).