Bug: Pass the dataset name/id for authorisation checks
The new authorisation checks need the dataset name/id for checking the
access level the user has on any data. This commit passes the dataset
name/id through the requests.

We also add some exceptions to ensure that this data is passed through
before even attempting to check for the authorisation.