Fix bug where mapping is sometimes wrongly submitted as a GET request
I think this bug was caused by changing the /run_mapping endpoint to
accept both GET and POST requests, and it only seemed to occur for some
users (only Rob that I'm aware of).

I'm still not sure why it was happening, but I addressed it by returning
/run_mapping to being only POST requests, and instead allowing the
mapping hashes to be linked to with the pattern /run_mapping/<hash>
(instead of the GET request /run_mapping?hash=<hash>). This appears to
have prevented the problem from happening.