BugFix: Use reworked `retrieve_phenotype_group_name` function
Use the reworked `retrieve_phenotype_group_name` to fetch the group
name for phenotype datasets.

== Bug Description ==

There was a subtle bug in the original calls to `retrieve_group_name`
function where the `dataset_id` was passed in as an argument, rather
than the group id.

For some Datasets, the `group_id` and `dataset_id` share the same
value, and would end up querying the correct values, but that was a
fluke where doing the wrong thing ended up with the expected values.

This commit fixes that by using the reworked function, which does the
correct indirection from `dataset_id` to the appropriate `group_id`.

See
https://github.com/genenetwork/genenetwork3/commit/088b6181457ef4e528b7dcbf04b7aca29b274d1f
for the reworked function.