Fix Bug: Unchanged Values Were Being Deleted
The original code (using the homebrew ORM system) would simply ignore
values that were `None` when doing updates - the new code using direct
queries was not.

My (fredmanglis) initial fix to provide a default for values in DB
that did not accept NULL was incorrect; instead, I needed to remove
any key-value pairs from the incoming data that were set to `None` to
fix the bug.

The fix still feels incorrect: maybe we should do direct comparisons
for all old-new value pairs, and set the data to be updated based on
the differences, rather than relying on `None`s.