Remove token and user detail handling from @app.before_request
The token and user details information is handled in the
`gn2.wqflask.oauth2.session`. Other parts of the system should make
use of that.

It also helps avoid some weird "action-at-a-distance" interactions -
this forces the code to request what it needs when it needs it and not
rely on some global variables.