Debug Haley-Knott Regression mapping failure
I suspect the use of relative paths might be the cause of the failure
in the production container — In the development environments and the
CI/CD container, the applications are run at the root of the GN2
repository/package. This might not be the case in the production-esque
containers, since the packages are installed in GNU Guix's store, and
the launching of the application might be entirely elsewhere.

This commit will hopefully log out the information necessary to figure
out whether or not my hypothesis is correct.