Improve the logic for checking if genotypes are databased
Previously it only checked if a genotype 'dataset' existed for the group in question.
Now it also checks if the first marker in the results is in the database. This isn't a perfect solution,
but it will probably avoid most situations where there's a mismatch between databased markers and markers in the genotype file