Avoid running code that gets other_samples for inapplicable groups
Previously there was a very time-consuming block of code checking which
samples with data in the DB were also part of the main samplelist
derived from the .geno file. But this code is completely unnecessary for
the vast majority of groups, which only have data for samples that are
included in the main .geno file

This is currently done via a variable set in webqtlConfig.py, just as a
stopgap. We should probably store this info somewhere else instead.