Remove mutual exclusion
Previously, the code had a mutual exclusion whose goal was to ensure
that only one thread could renew the token and all other threads would
then use that — at least, that was the idea. In reality, it did not
seem to work as expected, and even lead to an infinite loop.

This commit removes the mutual exclusion completely.