Explicitly ignore anything in the instance directory.
We ignore the files in the instance directory.

This directory can be used in development to hold a myriad of
generated files (e.g. performance profiling, caches, settings files,
etc). We do not want to track this files, since they are ephemeral by
nature, and even when not (as in the case of settings), they are
instance specific, therefore not useful for distributing.